Rename pyyed.py to __init__.py (resolve recursive import)
